Create myproject.c
Program a two-person game of Tic -Tac- Toe. The game is played on a three by
three board. Each player has a marker.
One player has an ‘X’, the other an ‘O’. Players alternate turns to place their marker
on the board. The first player to get three in a row either diagonally, horizontally, or
vertically, wins the games. In the event all squares are taken on the board without a
winner then it is a tie. The program should set up the game by asking for the names
of the players.
Player one should be assigned an ‘X’ as their marker, player two should be assigned
the ‘O’.
The players should then have the option to play again. You may not assume that any
input the user provides you is initially valid.
If the information provided by the user at any stage of the program is invalid, the
program should prompt until valid information is provided.